Refactors `MatterPhysics` methods into multiple files Adds stubs for `worker-matter.js`
