+++ mantra-game/plugins/physics-matter/MatterPhysics.js
@@ -3,6 +3,11 @@ import Matter from 'matter-js';
 import PhysicsInterface from './PhysicsInterface.js';
 import Collisions from '../collisions/Collisions.js';
 
+import collisionActive from './lib/collisionActive.js';
+import onAfterUpdate from './lib/onAfterUpdate.js';
+import setBodySize from './lib/setBodySize.js';
+import lockedProperties from './lib/lockedProperties.js';
+
 class MatterPhysics extends PhysicsInterface {
 
   static id = 'physics-matter';
@@ -18,18 +23,23 @@ class MatterPhysics extends PhysicsInterface {
     this.Bodies = Matter.Bodies;
     this.Composite = Matter.Composite;
     this.Events = Matter.Events;
+    this.useWorker = false;
 
     this.dimension = 2;
 
-    // TODO: add all collision events
+    this.Matter = Matter;
+
     //
     // collisionStart is used for initial collision detection ( like bullets or mines or player ship contact )
     //
-    //
     // collisionActive is used for continuous collision detection ( like touching a planet )
-    ////
+    //
     // collisionEnd is currently not being used
     //
+    this.collisionActive = collisionActive.bind(this);
+    this.onAfterUpdate = onAfterUpdate.bind(this);
+    this.setBodySize = setBodySize.bind(this);
+    this.lockedProperties = lockedProperties.bind(this);
 
   }
 
@@ -39,8 +49,51 @@ class MatterPhysics extends PhysicsInterface {
     this.engine.gravity.y = y;
   }
 
-  init(game) {
+  initWorkerMode() {
+    this.worker = new Worker('worker-matter.js');
+    this.worker.onmessage = this.handleWorkerMessage.bind(this);
 
+    // Send a message to the worker to initialize the engine
+    this.postMessageToWorker('initEngine', { config: this.config });
+  }
+
+  postMessageToWorker(action, data) {
+    this.worker.postMessage({ action, data });
+  }
+
+  handleWorkerMessage(event) {
+    const { action, ...messageData } = event.data;
+    switch (action) {
+      case 'engineInitialized':
+        console.log('Engine initialized in worker');
+        break;
+      case 'engineUpdated':
+        // Handle engine updated message, such as syncing state back to the main thread entities
+        break;
+      case 'bodyCreated':
+        console.log('Body created with ID:', messageData.bodyId);
+        // Update your main thread entity to reference the body ID, or other relevant data
+        break;
+      // Handle other messages...
+    }
+  }
+
+
+    
+
+  init(game) {
+    this.game = game;
+    if (this.useWorker) {
+      this.initWorkerMode();
+    } else {
+      this.initDirectMode();
+    }
+
+  }
+
+
+  initDirectMode() {
+    let game = this.game;
     this.engine = Matter.Engine.create()
     // game.systemsManager.addSystem('physics', this);
 
@@ -72,134 +125,10 @@ class MatterPhysics extends PhysicsInterface {
     // should this be onAfterUpdate? since we are serializing the state of the world?
     // i would assume we want that data *after* the update?
     let that = this;
-    this.onAfterUpdate(this.engine, function onAfterUpdate (event) {
-      // show total number of bodies in the world
-      // Remark: should this and bodyMap be replaced with a more generic mapping
-      // in order to allow non-physics backed entities to exist in the game?
-      for (const body of event.source.world.bodies) {
-
-        // let entity = that.game.getEntity(body.myEntityId);
-        let entity = that.game.data.ents._[body.myEntityId];
-
-        if (entity && body.isSleeping !== true && body.myEntityId) {
-
-          //
-          // Clamp max speed
-          //
-          let maxSpeed = 100; // TODO: move to config
-          if (entity.maxSpeed) {
-            maxSpeed = entity.maxSpeed;
-          }
-          limitSpeed(body, maxSpeed);
-
-          //
-          // Apply locked properties  ( like entity cannot move x or y position, etc )
-          //
-          if (entity.lockedProperties) {
-            that.lockedProperties(body);
-          }
-
-          // If this is the client and we are in online mode,
-          // do not update local physics for remote players, only update local physics for the local player
-          // This may be changed in the future or through configuration
-          if (that.game.isClient && that.game.onlineMode && entity.type === 'PLAYER' && entity.id !== game.currentPlayerId) {
-            // In online mode, if the entity is a player and that player is not the current player, skip updating physics
-            // continue; // Skip updating physics for remote players
-          }
-
-          if (that.game.isClient) {
-            // this is the logic for updating *all* entities positions
-            // this should probably be in entity-movement plugin
-            /*            */
-            // console.log(body.myEntityId, body.position)
-            let ent = that.game.entities.get(body.myEntityId);
-            // console.log('client ent', ent.id ,body.position)
-            // console.log('this.game.localGameLoopRunning', this.game.localGameLoopRunning)
-            if (that.game.localGameLoopRunning) {
-              // check if body position has changed
-
-              let bodyPosition = {
-                x: truncateToStringWithPrecision(body.position.x, 3),
-                y: truncateToStringWithPrecision(body.position.y, 3)
-              };
-              let entPosition = {
-                x: truncateToStringWithPrecision(ent.position.x, 3),
-                y: truncateToStringWithPrecision(ent.position.y, 3)
-              }
-
-              // TODO: add this same kind of logic for server as well?
-              // delta encoding will filter this; however it would be better to do it here as well
-              if (bodyPosition.x !== entPosition.x || bodyPosition.y !== entPosition.y) {
-                that.game.changedEntities.add(body.myEntityId);
-              }
-              // TODO: rotation / velocity as well, use flag isChanged
-
-              // check it z position is undefined on body ( 2D physics )
-              // if there is no z position, check for previous z position on entity and use that
-              // if there is no previous z position on entity, use 0
-
-              let position = { x: body.position.x, y: body.position.y };
-              if (typeof body.position.z === 'undefined') {
-                if (typeof ent.position.z === 'undefined') {
-                  position.z = 0;
-                } else {
-                  position.z = ent.position.z;
-                }
-              }
-
-              that.game.components.velocity.set(body.myEntityId, { x: body.velocity.x, y: body.velocity.y });
-              that.game.components.position.set(body.myEntityId, position);
-              that.game.components.rotation.set(body.myEntityId, body.angle);
-
-              // update size as well
-              //this.game.components.height.set(body.myEntityId, body.bounds.max.y);
-              //this.game.components.width.set(body.myEntityId, body.bounds.max.x);
-              // this.game.components.radius.set(body.myEntityId, body.bounds.max.x / 2);
-
-            }
-
-            if (body) {
-              if (that.game.systems.rbush) {
-                let ent = that.game.entities.get(body.myEntityId);
-                let position = { x: body.position.x, y: body.position.y };
-                that.game.systems.rbush.updateEntity({
-                  id: body.myEntityId,
-                  position: position,
-                  width: ent.width,
-                  height: ent.height
-                });
-                // this.game.deferredEntities[body.myEntityId] = ent;
-              }
-            }
-
-            if (ent.type === 'BULLET') {
-              that.game.changedEntities.add(body.myEntityId);
-              that.game.components.velocity.set(body.myEntityId, { x: body.velocity.x, y: body.velocity.y });
-              that.game.components.position.set(body.myEntityId, { x: body.position.x, y: body.position.y });
-              that.game.components.rotation.set(body.myEntityId, body.angle);
-            }
-
-          } else {
-            // this is the logic for updating *all* entities positions
-            // this should probably be in entity-movement plugin
-
-            let ent = that.game.getEntity(body.myEntityId);
-            //console.log('server ent', ent)
-            that.game.changedEntities.add(body.myEntityId);
-            that.game.components.velocity.set(body.myEntityId, { x: body.velocity.x, y: body.velocity.y });
-            that.game.components.position.set(body.myEntityId, { x: body.position.x, y: body.position.y });
-            that.game.components.rotation.set(body.myEntityId, body.angle);
-
-          }
-
-        }
-
-      }
-    });
+    this.onAfterUpdate(this.engine, this.onAfterUpdate);
 
     // TODO: configurable collision plugins
     game.use(new Collisions());
-
   }
 
   createBody(options) {
@@ -262,32 +191,6 @@ class MatterPhysics extends PhysicsInterface {
     Matter.Body.rotate(body, rotation);
   }
 
-  setBodySize(body, size) {
-    // size may have height and width ( rect ), or radius ( circle )
-    /*
-    if (typeof size.width !== 'undefined') {
-      Matter.Body.scale(body, size.width / body.bounds.max.x, 1);
-    }
-    if (typeof size.height !== 'undefined') {
-      Matter.Body.scale(body, 1, size.height / body.bounds.max.y);
-    }
-    */
-    if (typeof size.radius !== 'undefined') {
-      // Estimate the current radius as the average of width and height
-      let currentRadius = (body.bounds.max.x - body.bounds.min.x + body.bounds.max.y - body.bounds.min.y) / 4;
-
-      //console.log('Current Radius:', currentRadius);
-      //console.log('New Radius:', size.radius);
-
-      // Calculate the scale factor
-      let scaleFactor = size.radius / currentRadius;
-      //console.log('Scaling Factor:', scaleFactor);
-
-      // Apply the scaling
-      Matter.Body.scale(body, scaleFactor, scaleFactor);
-    }
-  }
-
   onBeforeUpdate(engine, callback) {
     Matter.Events.on(engine, 'beforeUpdate', callback);
   }
@@ -334,20 +237,6 @@ class MatterPhysics extends PhysicsInterface {
     });
   }
 
-  collisionActive(game, callback) {
-    Matter.Events.on(this.engine, 'collisionActive', (event) => {
-      for (let pair of event.pairs) {
-        const bodyA = pair.bodyA;
-        const bodyB = pair.bodyB;
-        // console.log('collisionActive', bodyA.entity, bodyB.entity)
-        if (bodyA.entity.collisionActive === true || bodyB.entity.collisionActive === true) {
-          game.emit('collision::active', { pair, bodyA, bodyB })
-          callback(pair, bodyA, bodyB);
-        }
-      }
-    });
-  }
-
   collisionEnd(game, callback) {
     Matter.Events.on(this.engine, 'collisionEnd', (event) => {
       for (let pair of event.pairs) {
@@ -375,23 +264,6 @@ class MatterPhysics extends PhysicsInterface {
     Matter.Events.off(this.engine, 'collisionEnd', callback);
   }
 
-  lockedProperties(body) {
-    let eId = body.myEntityId;
-    let ent = this.game.getEntity(eId);
-    if (ent && ent.lockedProperties) {
-      if (ent.lockedProperties.position) {
-        let currentPosition = body.position;
-        if (typeof ent.lockedProperties.position.x === 'number') {
-          currentPosition.x = ent.lockedProperties.position.x;
-        }
-        if (typeof ent.lockedProperties.position.y === 'number') {
-          currentPosition.y = ent.lockedProperties.position.y;
-        }
-        Matter.Body.setPosition(body, currentPosition);
-      }
-    }
-  }
-
 }
 
 export default MatterPhysics;
@@ -403,90 +275,6 @@ function limitSpeed(body, maxSpeed) {
     Matter.Body.setVelocity(body, newVelocity);
   }
 }
-
-
-
-/*
-
-class MatterPhysics extends PhysicsInterface {
-  constructor(workerMode = false) {
-    super();
-    if (workerMode) {
-      this.initWorkerMode();
-    } else {
-      this.initDirectMode();
-    }
-  }
-
-  initWorkerMode() {
-    this.worker = new Worker('matterWorker.js');
-    this.worker.onmessage = this.handleWorkerMessages.bind(this);
-    this.pendingRequests = new Map();
-    this.requestId = 0;
-
-    return new Proxy(this, {
-      get(target, prop) {
-        if (typeof target[prop] === 'function') {
-          return function (...args) {
-            return target.postMessageToWorker(prop, args);
-          }
-        }
-        return target[prop];
-      }
-    });
-  }
-
-  initDirectMode() {
-    // Here, directly bind Matter.js methods to this class
-    // for example, this.createEngine = Matter.Engine.create;
-    // and so on for all other methods
-  }
-
-}
-
-// 1:1 mapping use function, not proxy, simple call to object
-class MatterPhysics extends PhysicsInterface {
-  constructor() {
-    // ...
-
-    if (workerMode) {
-    if (isNode) {
-      // Node.js environment
-      const { fork } = require('child_process');
-      this.worker = fork('path/to/matterWorker.js');
-    } else {
-      // Browser environment
-      this.worker = new Worker('matterWorker.js');
-    }
-    this.worker.onmessage = this.handleWorkerMessages.bind(this);
-    // Additional setup...
-  } else {
-    this.initDirectMode();
-  }
-
-
-  }
-
-  handleWorkerMessages(event) {
-    if (event.data.command === 'engineUpdated') {
-      // Handle the updated physics state
-      // Update your game entities with the new physics data
-    }
-  }
-
-  createEngine(options) {
-    this.worker.postMessage({ command: 'createEngine', options });
-  }
-
-  updateEngine(delta) {
-    this.worker.postMessage({ command: 'updateEngine', delta });
-  }
-
-  // ... other methods
-}
-
-
-*/
 
 const truncateToStringWithPrecision = (value, precision = 3) => {
   return value.toFixed(precision);
